third_party/WebKit/LayoutTests/platform/win/virtual/gpu/fast/canvas/canvas-radial-gradient-spreadMethod-expected.txt

This is a testharness.js-based test.
FAIL Verify correct behaviour of spreadMethod to a radial gradient. assert_array_equals: property 0, expected 0 but got 1
Harness: the test ran to completion.
